Refactor header to it's own component
